feature-gate the datapath
