+++ Cargo.toml
@@ -4,23 +4,24 @@ version = "0.1.0"
 edition = "2021"
 
 [features]
-bin = ["structopt", "tracing-subscriber"]
-afd = []
+datapath = ["tun-tap", "libc", "socket2", "pnet", "flume"]
+bin = ["datapath", "clap", "tracing-subscriber"]
+afd = ["rand"]
 hwfq-audit = []
 
 [dependencies]
-tun-tap = { version = "0.1", default-features=false, features = [] }
-etherparse = "0.10"
-socket2 = { version = "0.5.3", features = ["all"] }
-pnet = "0.29"
-libc = "0.2"
-flume = "0.11"
-quanta = "0.9"
+quanta = "0.12"
 yaml-rust = "0.4"
 tracing = "0.1"
 color-eyre = "0.6"
-rand = "0.8"
-structopt = { version = "0.3", optional = true }
+etherparse = "0.14"
+tun-tap = { version = "0.1", default-features=false, features = [], optional = true }
+socket2 = { version = "0.5.3", features = ["all"], optional = true }
+pnet = { version = "0.29", optional = true }
+libc = { version = "0.2", optional = true }
+flume = { version = "0.11", optional = true }
+rand = { version = "0.8", optional = true }
+clap = { version = "4", features = ["derive"], optional = true }
 tracing-subscriber  = { version = "0.3", optional = true }
 
 [dev-dependencies]
@@ -28,6 +29,10 @@ tracing-subscriber = "0.3"
 
 [[bin]]
 name = "hwfq"
+required-features = ["bin", "afd"]
+
+[[bin]]
+name = "policer"
 required-features = ["bin"]
 
 [[bin]]
