+++ src/scheduler/afd.rs
@@ -221,44 +221,56 @@ mod t {
                 ip_hdr: etherparse::Ipv4Header::new(
                     100,
                     64,
-                    etherparse::IpNumber::Tcp,
+                    etherparse::IpNumber::TCP,
                     b_ip,
                     dst_ip,
-                ),
-                buf: vec![0u8; 100],
+                )
+                .unwrap(),
+                dport: 0,
+                fake_len: 100,
+                buf: vec![],
             })
             .unwrap();
             hwfq.enq(Pkt {
                 ip_hdr: etherparse::Ipv4Header::new(
                     100,
                     64,
-                    etherparse::IpNumber::Tcp,
+                    etherparse::IpNumber::TCP,
                     b_ip,
                     dst_ip,
-                ),
-                buf: vec![0u8; 100],
+                )
+                .unwrap(),
+                dport: 0,
+                fake_len: 100,
+                buf: vec![],
             })
             .unwrap();
             hwfq.enq(Pkt {
                 ip_hdr: etherparse::Ipv4Header::new(
                     100,
                     64,
-                    etherparse::IpNumber::Tcp,
+                    etherparse::IpNumber::TCP,
                     c_ip,
                     dst_ip,
-                ),
-                buf: vec![0u8; 100],
+                )
+                .unwrap(),
+                dport: 0,
+                fake_len: 100,
+                buf: vec![],
             })
             .unwrap();
             hwfq.enq(Pkt {
                 ip_hdr: etherparse::Ipv4Header::new(
                     100,
                     64,
-                    etherparse::IpNumber::Tcp,
+                    etherparse::IpNumber::TCP,
                     d_ip,
                     dst_ip,
-                ),
-                buf: vec![0u8; 100],
+                )
+                .unwrap(),
+                dport: 0,
+                fake_len: 100,
+                buf: vec![],
             })
             .unwrap();
 
@@ -299,7 +311,7 @@ mod t {
     #[test]
     fn afd_single_pair() {
         init();
-        let (mut hwfq, b_ip, c_ip, d_ip) = make_test_tree();
+        let (mut hwfq, b_ip, c_ip, _d_ip) = make_test_tree();
         let dst_ip = [42, 2, 0, 0];
         let mut b_cnt = 0;
         let mut c_cnt = 0;
@@ -311,11 +323,14 @@ mod t {
                     ip_hdr: etherparse::Ipv4Header::new(
                         100,
                         64,
-                        etherparse::IpNumber::Tcp,
+                        etherparse::IpNumber::TCP,
                         b_ip,
                         dst_ip,
-                    ),
-                    buf: vec![0u8; 100],
+                    )
+                    .unwrap(),
+                    dport: 0,
+                    fake_len: 100,
+                    buf: vec![],
                 })
                 .unwrap();
             }
@@ -323,11 +338,14 @@ mod t {
                 ip_hdr: etherparse::Ipv4Header::new(
                     100,
                     64,
-                    etherparse::IpNumber::Tcp,
+                    etherparse::IpNumber::TCP,
                     c_ip,
                     dst_ip,
-                ),
-                buf: vec![0u8; 100],
+                )
+                .unwrap(),
+                dport: 0,
+                fake_len: 100,
+                buf: vec![],
             })
             .unwrap();
